react/requirements.txt
flask
flask-cors
langchain
chromadb
huggingface-hub
transformers
groq
sentence-transformers
requests
python-dotenv
beautifulsoup4
langchainhub
langchain-community